+++ src/pages/MovieDetails/index.scss
@@ -9,6 +9,9 @@
   max-width: 500px;
   padding: 10px;
 }
+.details_container{
+  
+}
 
 .overview {
   display: flex;
